requirements.txt
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.16.1
beautifulsoup4==4.5.1
channels==0.17.3
constantly==15.1.0
daphne==0.15.0
dj-database-url==0.4.1
Django==1.10.3
django-autocomplete-light==3.2.1
django-autoslug==1.9.3
django-bootstrap3==7.1.0
django-lazysignup==1.1.1
-e git+https://github.com/jrief/django-websocket-redis@31723a7ae2de62a88c2b0d17950611052730f244#egg=django_websocket_redis
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
haikunator==2.1.0
incremental==16.10.1
msgpack-python==0.4.8
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
redis==2.10.5
requests==2.12.4
six==1.10.0
Twisted==16.5.0
txaio==2.5.2
whitenoise==3.2.2
wsaccel==0.6.2
zope.interface==4.3.2
